Switch `Size::from_px` constructor to `f32`
1f8e5db changed the `Size::as_px` method to return a floating point
value instead of an integer. This patch ensures the constructor is
consistent with that.

+++ src/lib.rs
@@ -121,8 +121,8 @@ impl Size {
     /// Create a new `Size` from px.
     ///
     /// The value will be clamped to the pt range of [`Size::new`].
-    pub fn from_px(size: u16) -> Self {
-        let pt = size as f32 * 72. / 96.;
+    pub fn from_px(size: f32) -> Self {
+        let pt = size * 72. / 96.;
         Size::new(pt)
     }
